+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This version includes **multiple breaking changes**. If you are upgrading from v
   - Sorting order issues when emails lack `receivedAt` headers.
 - IMAP: Fix `BINARY` fetch responses (#2940).
 - ACME: Allow requesting apex domain certificates.
+- Bootstrap: Accept RFC 6761 reserved TLDs during bootstrap.
 - Reverse proxy issues.
 - OSS builds.
 - DNS Updater: